orchestrator/requirements.txt
crewai>=0.80.0
crewai-tools>=0.17.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.0
langchain-ollama>=0.2.0
langchain-anthropic>=0.3.0
mem0ai>=0.1.0
qdrant-client>=1.11.0
pydantic>=2.8.0
